UefiPayloadPkg: Add support for Firmware Volume Block Protocol
Use coreboot's SMMSTORE v2 to implement a FVB protocol.
Add a library for SMMStore to be used in PEI and DXE.

The DXE component provides runtime services and takes care of virtual to physical
mapping the communication buffers between SMM and OS.

Make use of the APRIORI DXE to initialize an empty store on the first boot and
set the PCDs to sane values before the variable driver is loaded.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>